Refer to job definition by ARN
After specifying the job definition ARN in the respective IAM role, we
must refer to that exact same job definition revision at time of job
submission. This is done by referencing the job definition during
submission (in the manifest processor Lambda) by the same ARN provided
to the IAM role.